#3 #10 Moved sending responses logic to separate class. Made it run inside spring's executor's service
